requirements.txt
flask
flask-cors
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-migrate
psycopg2-binary
gunicorn